Changed handling of ba for matrix fetch slightly which got rid of fetch_active.

+++ branches/viceplus/v2.2-x64/vice/src/viciinew/vicii-chip-model.h
@@ -34,7 +34,7 @@ static inline BYTE get_sprite_ba_mask(unsigned int flags)
     return flags & 0xff;
 }
 
-static inline BYTE is_fetch_ba(unsigned int flags)
+static inline int is_fetch_ba(unsigned int flags)
 {
     return flags & 0x100;
 }
